LayoutTests/inspector/dom-debugger/event-animation-frame-breakpoints-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event AnimationFrame breakpoints.


== Running test suite: DOMDebugger.Event.AnimationFrame
-- Running test case: DOMDebugger.Event.AnimationFrame.AddBreakpoint
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleWindow_requestAnimationFrame
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] trigger_requestAnimationFrame
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.AddDisabledBreakpoint
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...
Disabling "animation-frame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.RemoveBreakpoint
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...
Removing "animation-frame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.RemoveDisabledBreakpoint
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...
Disabling "animation-frame" Event Breakpoint...
Removing "animation-frame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.Options.Condition
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.Options.IgnoreCount
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.Options.Action.Log
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.Options.Actions.Evaluate
Creating "animation-frame" Event Breakpoint...
Adding "animation-frame" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
-- Running test teardown.
